+++ lib/ui/screens/dashboard/tabs/control_tab_controller.dart
@@ -16,7 +16,7 @@ import 'package:mobileraker/data/dto/machine/output_pin.dart';
 import 'package:mobileraker/service/moonraker/printer_service.dart';
 import 'package:mobileraker/service/setting_service.dart';
 import 'package:mobileraker/service/ui/dialog_service.dart';
-import 'package:mobileraker/ui/components/dialog/edit_form/num_edit_form_viewmodel.dart';
+import 'package:mobileraker/ui/components/dialog/edit_form/num_edit_form_controller.dart';
 import 'package:mobileraker/ui/components/dialog/led_rgbw/led_rgbw_dialog_controller.dart';
 import 'package:mobileraker/ui/screens/dashboard/dashboard_controller.dart';
 import 'package:mobileraker/util/extensions/async_ext.dart';
